fix metrics

+++ generator/mcd_vae.py
@@ -343,7 +343,8 @@ plt.xlabel('epoch')
 plt.ylabel('loss')
 plt.title('KLD Loss')
 
-plt.savefig(metrics_path)
+if not os.path.exists(metrics_path):
+    plt.savefig(metrics_path)
 print("Done!!! Metrics of model is saved under name {}".format(NAME_METRICS+".png"))
 
 #generated data
